Move Backend to separate file
